requirements.txt
streamlit==1.43.2
numpy==2.2.4
fpdf==1.7.2
python-docx==1.1.2
